+++ src/Rpc/RpcServer.h
@@ -86,7 +86,7 @@ private:
   bool f_on_block_json(const F_COMMAND_RPC_GET_BLOCK_DETAILS::request& req, F_COMMAND_RPC_GET_BLOCK_DETAILS::response& res);
   bool f_on_transaction_json(const F_COMMAND_RPC_GET_TRANSACTION_DETAILS::request& req, F_COMMAND_RPC_GET_TRANSACTION_DETAILS::response& res);
   bool f_on_transactions_pool_json(const F_COMMAND_RPC_GET_POOL::request& req, F_COMMAND_RPC_GET_POOL::response& res);
-  bool f_getMixin(const Transaction& transaction, uint64_t& mixin);
+  bool f_getMixin(const Transaction& transaction, uint16_t& mixin);
 
   Logging::LoggerRef logger;
   core& m_core;
