+++ src/WalletLegacy/WalletSendTransactionContext.h
@@ -32,7 +32,7 @@ struct SendTransactionContext
   uint64_t foundMoney;
   std::vector<TransactionOutputInformation> selectedTransfers;
   TxDustPolicy dustPolicy;
-  uint64_t mixIn;
+  uint16_t mixIn;
   std::vector<tx_message_entry> messages;
   uint64_t ttl;
   uint32_t depositTerm;
